All emc related numbers now stored in Longs. Fixed bug with Novas (now you can break them without crash).

+++ src/main/java/moze_intel/projecte/emc/mappers/OreDictionaryMapper.java
@@ -11,14 +11,14 @@ import net.minecraftforge.oredict.OreDictionary;
 
 import java.util.Set;
 
-public class OreDictionaryMapper implements IEMCMapper<NormalizedSimpleStack, Integer> {
+public class OreDictionaryMapper implements IEMCMapper<NormalizedSimpleStack, Long> {
 
 	private static final Set<String> BLACKLIST_EXCEPTIONS = Sets.newHashSet(
 		"dustPlastic"
 	);
 
 	@Override
-	public void addMappings(IMappingCollector<NormalizedSimpleStack, Integer> mapper, Configuration config) {
+	public void addMappings(IMappingCollector<NormalizedSimpleStack, Long> mapper, Configuration config) {
 		if (config.getBoolean("blacklistOresAndDusts", "", true, "Set EMC=0 for everything that has an OD Name that starts with `ore`, `dust` or `crushed` besides `dustPlastic`")) {
 			//Black-list all ores/dusts
 			for (String s : OreDictionary.getOreNames()) {
@@ -38,8 +38,8 @@ public class OreDictionaryMapper implements IEMCMapper<NormalizedSimpleStack, In
 							continue;
 						}
 
-						mapper.setValueBefore(NSSItem.create(stack), 0);
-						mapper.setValueAfter(NSSItem.create(stack), 0);
+						mapper.setValueBefore(NSSItem.create(stack), 0l);
+						mapper.setValueAfter(NSSItem.create(stack), 0l);
 					}
 				}
 			}
